+++ lib/algorithm/pitch_algorithm.dart
@@ -1,4 +1,4 @@
-import 'package:yin/pitch_detector_result.dart';
+import 'package:pitch_detector_dart/pitch_detector_result.dart';
 
 abstract class PitchAlgorithm {
   PitchDetectorResult getPitch(final List<double> audioBuffer);