+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* First release of the dart YIN library.
+* First release of the pitch detector dart library.